requirements.txt
pytorch 
resemblyzer
